Update README.md
This edited description better encompasses the project's mission of focusing on the individual who is listening....since frenchy is creating the template, this first filtered curation database store is from frenchy's perspective.  In the future different people will have their own filtered list using this template.  For now, frenchy has to focus on frenchy's historic material because 1) frenchy must gather the historical footage to create a coherent repository and 2) it is impossible to store all other peoples database here also, so lets focus on frenchy for now.  In the future, it would be awesome to filter for each individual of the universe and present the jamography of their desires....which uses this template.